Remove Stage Subset (to refactoring to process group concept)
